Move core hitcount logic to parent abstract class
